# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Contacts API",
    "description": "A comprehensive REST API for managing contacts with full CRUD operations. Built for CSE 341 Web Services.",
    "version": "1.0.0",
    "contact": {
      "name": "API Support",
      "url": "https://github.com/your-repo",
      "email": "support@contactsapi.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:3001",
  "basePath": "/",
  "tags": [
    {
      "name": "Contacts",
      "description": "Contact management endpoints"
    },
    {
      "name": "Health",
      "description": "Server health check"
    }
  ],
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
    "/": {
      "get": {
        "tags": [],
        "summary": "API Welcome",
        "description": "Root endpoint with API information",
        "responses": {
          "200": {
            "description": "API information",
            "schema": {
              "type": "object",
              "properties": {
                "message": { "type": "string" },
                "description": { "type": "string" },
                "version": { "type": "string" },
                "endpoints": { "type": "object" }
              }
            }
          }
        }
      }
    },
    "/contacts": {
      "get": {
        "tags": ["Contacts"],
        "summary": "Get all contacts",
        "description": "Retrieve a list of all contacts",
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Contact"
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "post": {
        "tags": ["Contacts"],
        "summary": "Create a new contact",
        "description": "Create a new contact with the provided information",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ContactInput"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Contact created successfully",
            "schema": {
              "$ref": "#/definitions/Success"
            }
          },
          "400": {
            "description": "Bad Request - Invalid input",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/contacts/{id}": {
      "get": {
        "tags": ["Contacts"],
        "summary": "Get contact by ID",
        "description": "Retrieve a single contact by its ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "MongoDB ObjectId of the contact"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/Contact"
            }
          },
          "400": {
            "description": "Invalid ID format",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "404": {
            "description": "Contact not found",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "put": {
        "tags": ["Contacts"],
        "summary": "Update a contact",
        "description": "Update an existing contact by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "MongoDB ObjectId of the contact"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ContactInput"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Contact updated successfully",
            "schema": {
              "$ref": "#/definitions/Success"
            }
          },
          "400": {
            "description": "Invalid input data or ID format",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "404": {
            "description": "Contact not found",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "delete": {
        "tags": ["Contacts"],
        "summary": "Delete a contact",
        "description": "Delete a contact by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "MongoDB ObjectId of the contact"
          }
        ],
        "responses": {
          "200": {
            "description": "Contact deleted successfully",
            "schema": {
              "$ref": "#/definitions/Success"
            }
          },
          "400": {
            "description": "Invalid ID format",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "404": {
            "description": "Contact not found",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "tags": ["Health"],
        "summary": "Health check",
        "description": "Check if the API is running",
        "responses": {
          "200": {
            "description": "API is healthy and running",
            "schema": {
              "type": "object",
              "properties": {
                "status": { "type": "string", "example": "OK" },
                "message": { "type": "string", "example": "Contacts API is running" },
                "timestamp": { "type": "string", "format": "date-time" },
                "environment": { "type": "string", "example": "development" },
                "version": { "type": "string", "example": "1.0.0" }
              }
            }
          }
        }
      }
    }
  },
  "definitions": {
    "Contact": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string",
          "description": "Auto-generated MongoDB ObjectId",
          "example": "507f1f77bcf86cd799439011"
        },
        "firstName": {
          "type": "string",
          "description": "First name of the contact",
          "example": "John"
        },
        "lastName": {
          "type": "string",
          "description": "Last name of the contact",
          "example": "Doe"
        },
        "email": {
          "type": "string",
          "format": "email",
          "description": "Email address of the contact",
          "example": "john.doe@example.com"
        },
        "favoriteColor": {
          "type": "string",
          "description": "Favorite color of the contact",
          "example": "Blue"
        },
        "birthday": {
          "type": "string",
          "format": "date",
          "description": "Birthday of the contact in YYYY-MM-DD format",
          "example": "1990-05-15"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time",
          "description": "Timestamp when the contact was created"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time",
          "description": "Timestamp when the contact was last updated"
        }
      }
    },
    "ContactInput": {
      "type": "object",
      "required": ["firstName", "lastName", "email", "favoriteColor", "birthday"],
      "properties": {
        "firstName": {
          "type": "string",
          "description": "First name (2-50 characters)",
          "example": "John",
          "minLength": 2,
          "maxLength": 50
        },
        "lastName": {
          "type": "string",
          "description": "Last name (2-50 characters)",
          "example": "Doe",
          "minLength": 2,
          "maxLength": 50
        },
        "email": {
          "type": "string",
          "format": "email",
          "description": "Valid email address",
          "example": "john.doe@example.com"
        },
        "favoriteColor": {
          "type": "string",
          "description": "Favorite color",
          "example": "Blue"
        },
        "birthday": {
          "type": "string",
          "format": "date",
          "description": "Birthday in YYYY-MM-DD format",
          "example": "1990-05-15"
        }
      }
    },
    "Error": {
      "type": "object",
      "properties": {
        "error": {
          "type": "string",
          "description": "Error message"
        },
        "details": {
          "type": "string",
          "description": "Additional error details"
        }
      }
    },
    "Success": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "description": "Success message"
        },
        "_id": {
          "type": "string",
          "description": "Contact ID when applicable"
        }
      }
    }
  }
}